mastodon: fix package so test passes

+++ pkgs/servers/mastodon/default.nix
@@ -1,5 +1,5 @@
 { nodejs-slim, yarn2nix-moretea, fetchFromGitHub, bundlerEnv,
-  stdenv, yarn, lib, callPackage, ... }:
+  stdenv, yarn, imagemagick, ffmpeg, file, lib, callPackage, ... }:
 
 let
   version = import ./version.nix;
@@ -52,10 +52,13 @@ in stdenv.mkDerivation {
     ln -s ${mastodon-assets}/public/assets public/assets
     ln -s ${mastodon-assets}/public/packs public/packs
 
+    sed -i 's#/usr/bin/env ruby#${mastodon-gems.wrappedRuby}/bin/ruby#g' bin/*
+
     for b in $(ls ${mastodon-gems}/bin/)
     do
-      rm -f bin/$b
-      ln -s ${mastodon-gems}/bin/$b bin/$b
+      if [ ! -f bin/$b ]; then
+        ln -s ${mastodon-gems}/bin/$b bin/$b
+      fi
     done
 
     rm -rf log
